allow installation to external media (microSD)
This change will enable installation of the app to external as well as internal flash memory which is important for devices where internal memory is limited but external memory is cheap and abundant in form of micro SD cards. Please have a look at https://developer.android.com/guide/topics/data/install-location for further information. Thank you for your consideration.